% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{beautify_rule}
\alias{beautify_rule}
\title{Replaces consecutive "&"s and "|"s in a string to a single & and |.}
\usage{
beautify_rule(rule)
}
\arguments{
\item{rule}{A string containing positive integers, parentheses, and \code{"&"} and \code{"|"} only.}
}
\value{
A string with extra \code{"&"}s and \code{"|"}s removed.
}
\description{
Replaces consecutive \code{"&"}s and \code{"|"}s in a string to a single \code{"&"} and \code{"|"}.
}
\details{
Applied to \code{domain$rule} if \code{domain$type == "polynomial"}.
}
\examples{
beautify_rule("(1 & 2 && 3 &&& 4) | 5 || 6 ||| 7")
}
