% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorizeSamples.R
\name{categorizeSamples}
\alias{categorizeSamples}
\title{Categorize samples on groups}
\usage{
categorizeSamples(scores, lowThreshold = 0.25, highThreshold = 0.75)
}
\arguments{
\item{scores}{array of collapsed scores.}

\item{lowThrehsold}{Threshold within samples are considered in "low" group.}

\item{highThrehsold}{Threshold within samples are considered in "high" group.}
}
\value{
Array with categorized values in three groups. This array have three different values
         "-1" samples are categorized at "low" threshold
         "0" samples are categorized at "middle" threshold
         "1" samples are categorized at "high" threshold
}
\description{
Categorize samples on groups
}
\examples{
categorizeSamples(c(1,2,3,4))
categorizeSamples(c(1,2,3,4), 0.25, 0.75) 

}
\author{
Luis Palomero, \email{lpalomerol@gmail.com}

This method categorizes samples in left-right groups
}
