% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_init.R
\name{compare_two_sub_results}
\alias{compare_two_sub_results}
\title{Compares two lists returned from get_results().}
\usage{
compare_two_sub_results(res, res2)
}
\arguments{
\item{res}{A res list returned from \code{get_results()}.}

\item{res2}{A res list returned from \code{get_results()}.}
}
\value{
A list of numbers all of which should be close to 0 if \code{res} and \code{res2} are expected to be the same.
}
\description{
Compares two lists returned from \code{get_results()}.
}
