% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{crbound_mu}
\alias{crbound_mu}
\title{The Cramér-Rao lower bound (times \code{n}) for estimating the mean parameter from a univariate truncated normal sample with known variance parameter.}
\usage{
crbound_mu(mu, sigmasq)
}
\arguments{
\item{mu}{The mean parameter.}

\item{sigmasq}{The variance parameter.}
}
\value{
A number, the Cramér-Rao lower bound.
}
\description{
The Cramér-Rao lower bound (times \code{n}) on the variance for estimating the mean parameter \code{mu} from a univariate truncated normal sample, assuming the true variance parameter \code{sigmasq} is known.
}
\details{
The Cramér-Rao lower bound in this case is defined as \eqn{\sigma^4/var(X-\mu)}.
}
