% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_init.R
\name{diff_lists}
\alias{diff_lists}
\title{Computes the sum of absolute differences between two lists.}
\usage{
diff_lists(l1, l2, name = NULL)
}
\arguments{
\item{l1}{A list.}

\item{l2}{A list.}

\item{name}{A string, default to \code{NULL}. If not \code{NULL}, computes the differences in the \code{l1[[name]]} and \code{l2[[name]]}.}
}
\value{
Returns the sum of absolute differences between \code{l1} and \code{l2} if \code{name} is \code{NULL}, or that between \code{l1[[name]]} and \code{l2[[name]]} otherwise. If \code{name} is not \code{NULL} and if \code{name} is in exactly one of \code{l1} and \code{l2}, returns \code{Inf}; if \code{name} is in neither, returns \code{NA}. Exception: Returns a positive integer if the two elements compared hold \code{NA}, \code{NULL} or \code{Inf} values in different places.
}
\description{
Computes the sum of absolute differences between two lists using diff_vecs().
}
