% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_init.R
\name{diff_vecs}
\alias{diff_vecs}
\title{Computes the sum of absolute differences in the finite non-NA/NULL elements between two vectors.}
\usage{
diff_vecs(l1, l2, relative = FALSE)
}
\arguments{
\item{l1}{A vector.}

\item{l2}{A vector.}

\item{relative}{A boolean, default to \code{FALSE}. If \code{TRUE}, returns the relative difference (sum of absolute differences divided by the elementwise minimum between \code{l1} and \code{l2}).}
}
\value{
The sum of (relative) absolute differences in \code{l1} and \code{l2}, or a positive integer if two vectors differ in length or hold \code{NA}, \code{NULL} or \code{Inf} values in different places.
}
\description{
Computes the sum of absolute differences in the finite non-\code{NA}/\code{NULL} elements between two vectors.
}
