% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{gcd}
\alias{gcd}
\title{Finds the greatest (positive) common divisor of two integers.}
\usage{
gcd(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
The greatest (positive) common divisor of two integers; if one of them is 0, returns the absolute value of the other number.
}
\description{
Finds the greatest (positive) common divisor of two integers; if one of them is 0, returns the absolute value of the other number.
}
\examples{
gcd(1, 2)
gcd(1, -2)
gcd(12, -18)
gcd(-12, 18)
gcd(15, 0)
gcd(0, -15)
gcd(0, 0)
}
