% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genScore.R
\name{genScore}
\alias{genScore}
\title{Genetic Score metric}
\usage{
genScore(varX, varY)
}
\arguments{
\item{varX}{array of "X" dimension values.}

\item{varY}{array of "Y" dimension values}
}
\value{
Array with the antisense distance metric, as more greater is this value
        the point is more far than (0,x) line, that is, more on the right corner of the plot.
}
\description{
Genetic Score metric
}
\examples{
genScore(c(1,2,3,4), c(4,3,2,1))

}
\author{
Luis Palomero, \email{lpalomerol@gmail.com}

This metric "collapses" two-dimensional values into a one dimensional. 
This distance is related to the inverse diagonal from a 2D space metric varA/varB
}
