% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{get_crit_nopenalty}
\alias{get_crit_nopenalty}
\title{Minimized loss for unpenalized restricted asymmetric models.}
\usage{
get_crit_nopenalty(elts, exclude = NULL, exclude_eta = NULL,
  previous_res = NULL)
}
\arguments{
\item{elts}{A list, elements necessary for calculations returned by get_elts().}

\item{exclude}{Optional. A p*p binary matrix or a p^2 binary vector, where \code{1} indicates the entry in K was estimated to 0 in the previous estimate. Default to \code{NULL}.}

\item{exclude_eta}{Optional. A p-binary vector, similar to \code{exclude}. Default to \code{NULL}.}

\item{previous_res}{Optional. A list, the returned list by \code{get_results()} run previously with another lambda value. Default to \code{NULL}.}
}
\value{
A number, the refitted loss.
}
\description{
Analytic solution of the minimized loss for an unpenalized asymmetric model restricted to a given support. Does not work if \code{symmetric == "symmetric"}.
}
\details{
If \code{previous_res} is provided, \code{exclude} and \code{exclude_eta} must be \code{NULL} or be consistent with the estimated support in \code{previous_res}. If \code{previous_res} and \code{exclude} are both \code{NULL}, assume all edges are present. The same applies to the non-profiled non-centered case when \code{previous_res} and \code{exclude_eta} are both \code{NULL}.
}
\examples{
if (!requireNamespace("tmvtnorm", quietly = TRUE)){
  stop("Please install package \\"tmvtnorm\\" first.", call. = FALSE)
}
require(tmvtnorm)
n <- 50
p <- 30
h_hp <- get_h_hp("min_pow", 1, 3)
mu <- rep(0, p)
K <- diag(p)
dm <- 1 + (1-1/(1+4*exp(1)*max(6*log(p)/n, sqrt(6*log(p)/n))))
x <- tmvtnorm::rtmvnorm(n, mean = mu, sigma = solve(K),
       lower = rep(0, p), upper = rep(Inf, p), algorithm = "gibbs",
       burn.in.samples = 100, thinning = 10)

elts_NC_NP <- get_elts(h_hp$h, h_hp$hp, x, setting="trun_gaussian",
                centered=FALSE, profiled=FALSE, diag=dm)
res_nc_np <- get_results(elts_NC_NP, symmetric="symmetric", lambda1=0.35,
               lambda2=2, previous_res=NULL, is_refit=FALSE)
get_crit_nopenalty(elts_NC_NP, previous_res=res_nc_np)
}
