% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{get_safe_log_h_hp}
\alias{get_safe_log_h_hp}
\title{Asymptotic log of \code{h} and \code{hp} functions for large \code{x} for some modes.}
\usage{
get_safe_log_h_hp(mode, para)
}
\arguments{
\item{mode}{A string, the class of the \code{h} function.}

\item{para}{A number, the first parameter to the \code{h} function.}
}
\value{
A list of two functions, \code{logh} and \code{loghp}.
}
\description{
Asymptotic log of \code{h} and \code{hp} functions for large \code{x} for some modes.
}
