% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{makecoprime}
\alias{makecoprime}
\title{Makes two integers coprime.}
\usage{
makecoprime(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
The greatest (positive) common divisor of two integers; if one of them is 0, returns the absolute value of the other number.
}
\description{
Divides both integers by their greatest common divisor, switching their signs if the second integer is negative. If either integer is 0, returns without modification.
}
\examples{
makecoprime(1, 2)
makecoprime(1, -2)
makecoprime(12, -18)
makecoprime(-12, 18)
makecoprime(15, 0)
makecoprime(0, -15)
makecoprime(0, 0)
}
