% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{rab_arms_R}
\alias{rab_arms_R}
\title{Random data generator from general \code{a}-\code{b} distributions.}
\usage{
rab_arms_R(n, a, b, eta, K, seed = NULL, burn_in = 1000,
  thinning = 1000, verbose = TRUE)
}
\arguments{
\item{n}{A number, number of observations.}

\item{a}{A number, must be strictly larger than \eqn{b/2}.}

\item{b}{A number, must be >= 0.}

\item{eta}{A vector, the linear part in the distribution.}

\item{K}{A strictly co-positive square matrix, the interaction matrix.}

\item{seed}{Optional. A number, the seed for the random generator.}

\item{burn_in}{Optional. A positive integer, the number of burn-in samples in ARMS to be discarded.}

\item{thinning}{Optional. A positive integer, thinning factor in ARMS. Samples are taken at iteration steps \eqn{\mathrm{burn\_in}+1}{burn_in+1}, \eqn{\mathrm{burn\_in}+1+\mathrm{thinning}}{burn_in+1+thinning}, ..., \eqn{\mathrm{burn\_in}+1+(n-1)*\mathrm{thinning}}{burn_in+1+(n-1)*thinning}. Default to \code{1000}.}

\item{verbose}{Optional. A boolean. If \code{TRUE}, prints a progress bar showing the progress. Defaults to \code{TRUE}.}
}
\value{
An \eqn{n\times p}{n*p} matrix of samples, where \eqn{p} is the length of \code{eta}.
}
\description{
Random data generator from general \code{a}-\code{b} graphs using adaptive rejection metropolis sampling (ARMS).
}
\details{
Randomly generates \code{n} samples from the \code{p}-variate \code{a}-\code{b} distributions with parameters \eqn{\boldsymbol{\eta}}{\eta} and \eqn{\mathbf{K}}{K}, where \code{p} is the length of \eqn{\boldsymbol{\eta}}{\eta} or the dimension of the square matrix \eqn{\mathbf{K}}{K}.

The \code{a}-\code{b} distribution is proportional to
\deqn{\exp\left(-\frac{1}{2a}{\boldsymbol{x}^a}^{\top}\mathbf{K}{\boldsymbol{x}}^a+\boldsymbol{\eta}^{\top}\frac{\boldsymbol{x}^b-\mathbf{1}_p}{b}\right)}{exp(-x^a'Kx^a/(2a)+eta'(x^b-rep(1,p))/b)}
on the non-negative orthant of \eqn{\mathbf{R}^p}{R^p}.
}
\examples{
p <- 30
set.seed(1)
eta <- stats::rnorm(p)*0.5
K <- cov_cons("er", p, seed = 3, spars = 0.05, eig = 0.1)
rab_arms_R(n=100, a=1.5, b=2.3, eta=eta, K=K,
  seed=1, burn_in=100, thinning=1000, verbose=TRUE)
rab_arms_R(n=100, a=1.2, b=0.9, eta=eta, K=K,
  seed=2, burn_in=200, thinning=2000, verbose=TRUE)
}
