% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{refit}
\alias{refit}
\title{Loss for a refitted (restricted) unpenalized model}
\usage{
refit(res, elts)
}
\arguments{
\item{res}{A list of results returned by \code{get_results()}.}

\item{elts}{A list, elements necessary for calculations returned by \code{get_elts()}.}
}
\value{
A number, the loss of the refitted model.
}
\description{
Refits an unpenalized model restricted to the estimated edges, with \code{lambda1=0}, \code{lambda2=0} and \code{diagonal_multiplier=1}. Returns \code{Inf} if no unique solution exists (when the loss is unbounded from below or has infinitely many minimizers).
}
\details{
Currently the function only returns \code{Inf} when the maximum node degree is >= the sample size, which is a sufficient and necessary condition for inexistence of a unique solution with probability 1 if \code{symmetric != "symmetric"}. In practice it is also a sufficient and necessary condition for most cases and \code{symmetric == "symmetric"}.
}
\examples{
if (!requireNamespace("tmvtnorm", quietly = TRUE)){
  stop("Please install package \\"tmvtnorm\\" first.", call. = FALSE)
}
require(tmvtnorm)
n <- 50
p <- 30
h_hp <- get_h_hp("min_pow", 1, 3)
mu <- rep(0, p)
K <- diag(p)
dm <- 1 + (1-1/(1+4*exp(1)*max(6*log(p)/n, sqrt(6*log(p)/n))))
x <- tmvtnorm::rtmvnorm(n, mean = mu, sigma = solve(K),
       lower = rep(0, p), upper = rep(Inf, p), algorithm = "gibbs",
       burn.in.samples = 100, thinning = 10)

elts_NC_NP <- get_elts(h_hp$h, h_hp$hp, x, setting="trun_gaussian",
                centered=FALSE, profiled=FALSE, diag=dm)
res_nc_np <- get_results(elts_NC_NP, symmetric="symmetric",
               lambda1=0.35, lambda2=2, previous_res=NULL, is_refit=FALSE)
refit(res_nc_np, elts_NC_NP)

elts_NC_P <- get_elts(h_hp$h, h_hp$hp, x, setting="trun_gaussian",
               centered=FALSE, profiled=TRUE, diag=dm)
res_nc_p <- get_results(elts_NC_P, symmetric="symmetric",
              lambda1=0.35, lambda2=NULL, previous_res=NULL, is_refit=FALSE)
refit(res_nc_p, elts_NC_P)

elts_C <- get_elts(h_hp$h, h_hp$hp, x, setting="trun_gaussian",
            centered=TRUE, diag=dm)
res_c <- get_results(elts_C, symmetric="or", lambda1=0.35,
           lambda2=NULL, previous_res=NULL, is_refit=FALSE)
refit(res_c, elts_C)

}
