% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{rexp_gamma_reject_R}
\alias{rexp_gamma_reject_R}
\title{Random data generator from exp or gamma graphs.}
\usage{
rexp_gamma_reject_R(n, gamm, eta, K, seed = NULL, burn_in = 1000,
  thinning = 1000, max_iter = 1e+05, verbose = TRUE)
}
\arguments{
\item{n}{A number, number of observations.}

\item{gamm}{A boolean, generate from gamma (\code{TRUE}) or exponential square-root (\code{FALSE}) graphs.}

\item{eta}{A vector, the linear part in the distribution.}

\item{K}{A strictly co-positive square matrix, the interaction matrix.}

\item{seed}{Optional. A number, the seed for the random generator.}

\item{burn_in}{Optional. A positive integer, the number of burn-in samples in Gibbs sampling to be discarded.}

\item{thinning}{Optional. A positive integer, thinning factor in Gibbs sampling. Samples are taken at iteration steps \eqn{\mathrm{burn\_in}+1}{burn_in+1}, \eqn{\mathrm{burn\_in}+1+\mathrm{thinning}}{burn_in+1+thinning}, ..., \eqn{\mathrm{burn\_in}+1+(n-1)*\mathrm{thinning}}{burn_in+1+(n-1)*thinning}. Default to \code{1000}.}

\item{max_iter}{Optional. A positive integer, the maximum number of proposals for each sample. Default to \code{100000}.}

\item{verbose}{Optional. A boolean. If \code{TRUE}, prints a progress bar showing the progress. Defaults to \code{TRUE}.}
}
\value{
An \eqn{n\times p}{n*p} matrix of samples, where \eqn{p} is the length of \code{eta}.
}
\description{
Random data generator from exponential square-root (\eqn{a=0.5}, \eqn{b=0.5}) or gamma (\eqn{a=0.5}, \eqn{b=0}) graphs using basic rejection sampling.
}
\details{
Note: rab_arms_R() with a=0.5 and b=0.5 (exp) or b=0 (gamma) may be a more stable generator.

Randomly generates \code{n} samples from the \code{p}-variate exponential square-root or gamma distributions with parameters \code{eta} and \code{K}, where \code{p} is the length of \code{eta} or the dimension of the square matrix \code{K}.

The exponential square-root distribution is proportional to
\deqn{\exp\left(-\sqrt{\boldsymbol{x}}^{\top}\mathbf{K}\sqrt{\boldsymbol{x}}+2\boldsymbol{\eta}^{\top}\sqrt{\boldsymbol{x}}\right)}{exp(-sqrt(x)'K*sqrt(x)+2*\eta'sqrt(x))}
on the non-negative orthant of \eqn{\mathbf{R}^p}{R^p}.

The gamma distribution is proportional to
\deqn{\exp\left(-\sqrt{\boldsymbol{x}}^{\top}\mathbf{K}\sqrt{\boldsymbol{x}}+\boldsymbol{\eta}^{\top}\log(\boldsymbol{x})\right)}{exp(-sqrt(x)'K*sqrt(x)+\eta'log(x))}
on the non-negative orthant of \eqn{\mathbf{R}^p}{R^p}.
}
\examples{
p <- 30
set.seed(1)
eta <- stats::rnorm(p)*0.5
eta[eta <= -1] <- abs(eta[eta <= -1])
K <- cov_cons("er", p, seed = 3, spars = 0.05, eig = 0.1)
rexp_gamma_reject_R(n=1000, gamm=TRUE, eta=eta, K=K, seed=1,
  burn_in=100, thinning=1000, max_iter=1e+05, verbose=TRUE)
rexp_gamma_reject_R(n=1000, gamm=FALSE, eta=eta, K=K, seed=2,
  burn_in=200, thinning=2000, max_iter=1e+05, verbose=TRUE)
}
