% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{s_at}
\alias{s_at}
\title{Returns the character at a position of a string.}
\usage{
s_at(string, position)
}
\arguments{
\item{string}{A string.}

\item{position}{A positive number.}
}
\value{
A character
}
\description{
Returns the character at a position of a string.
}
\details{
Calls \code{substr(string, position, position)}.
}
\examples{
s_at("123", 1)
s_at("123", 2)
s_at("123", 3)
s_at("123", 4)
s_at("123", 0)
}
