% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{s_output}
\alias{s_output}
\title{Helper function for outputting if verbose.}
\usage{
s_output(out, verbose, verbosetext)
}
\arguments{
\item{out}{Text string.}

\item{verbose}{Boolean.}

\item{verbosetext}{Text string.}
}
\value{
If \code{verbose == TRUE}, outputs a string that concatenates \code{verbosetext} and \code{out}.
}
\description{
Helper function for outputting if verbose.
}
