% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{eBIC}
\alias{eBIC}
\title{eBIC score with or without refitting.}
\usage{
eBIC(res, elts, BIC_refit = TRUE, gammas = c(0, 0.5, 1))
}
\arguments{
\item{res}{A list of results returned by get_results().}

\item{elts}{A list, elements necessary for calculations returned by get_elts().}

\item{BIC_refit}{A boolean, whether to get the BIC scores by refitting an unpenalized model restricted to the estimated edges, with \code{lambda1=0}, \code{lambda2=0} and \code{diagonal_multiplier=1}. Default to \code{TRUE}.}

\item{gammas}{Optional. A number of a vector of numbers. The \eqn{\gamma} parameter in eBIC. Default to \code{c(0,0.5,1)}.}
}
\value{
A vector of length \code{2*length(gammas)}. The first \code{length(gammas)} numbers are the eBIC scores without refitting for each \code{gamma} value, and the rest are those with refitting if \code{BIC_refit == TRUE}, or \code{Inf} if \code{BIC_refit == FALSE}.
}
\description{
Calculates the eBIC score both with and without refitting an unpenalized model restricted to the estimated support.
}
\examples{
# Examples are shown for Gaussian truncated to R+^p only. For other distributions
#   on other types of domains, please refer to \code{gen()} or \code{get_elts()}, 
#   as the way to call this function (\code{eBIC()}) is exactly the same in those cases.
if (!requireNamespace("tmvtnorm", quietly = TRUE))
   stop("Please install package \"tmvtnorm\".")
n <- 50
p <- 30
domain <- make_domain("R+", p=p)
h_hp <- get_h_hp("min_pow", 1, 3)
mu <- rep(0, p)
K <- diag(p)
dm <- 1 + (1-1/(1+4*exp(1)*max(6*log(p)/n, sqrt(6*log(p)/n))))
x <- tmvtnorm::rtmvnorm(n, mean = mu, sigma = solve(K),
       lower = rep(0, p), upper = rep(Inf, p), algorithm = "gibbs",
       burn.in.samples = 100, thinning = 10)

elts_gauss_np <- get_elts(h_hp, x, setting="gaussian", domain=domain,
                centered=FALSE, profiled=FALSE, diag=dm)
res_nc_np <- get_results(elts_gauss_np, symmetric="symmetric",
               lambda1=0.35, lambda2=2, previous_res=NULL,
               is_refit=FALSE)
eBIC(res_nc_np, elts_gauss_np, BIC_refit=TRUE, gammas=c(0,0.5,1))
}
