% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genScore.R
\name{genScore}
\alias{genScore}
\title{Genetic Score metric}
\usage{
genScore(varA, varB)
}
\arguments{
\item{varA}{array of "X" dimension values.}

\item{varB}{array of "Y" dimension values}
}
\value{
Array with the antisense distance metric
}
\description{
This metric "collapses" two-dimensional values into a one dimensional. 
This distance is related to the inverse diagonal from a 2D space metric varA/varB
}
\examples{
genScore(c(1,2,3,4), c(4,3,2,1))

}
