% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscore.R
\name{get_trun}
\alias{get_trun}
\title{The truncation point for \code{h}.}
\usage{
get_trun(mode, param1, param2)
}
\arguments{
\item{mode}{A string, the class of the \code{h} function.}

\item{param1}{A number, the first parameter to the \code{h} function.}

\item{param2}{A number, the second parameter (may be optional depending on \code{mode}) to the \code{h} function.}
}
\value{
Returns the truncation point (the point where \code{h} becomes constant and \code{hp} becomes 0) for some selected modes.
}
\description{
The truncation point for \code{h}.
}
